feat: intialize custom config
feat: label keys and layers with emoji

fix: use MOD_RALT instead of RALT

RALT and LALT must refer to the same keycode, which caused the
mod-morphs `&left_home` and `&right_end` to be triggered by both keys.
This commit changes the mod-morphs to use MOD_RALT instead of RALT,
which only refers to the right alt key.